Code encapsulation
Added a new class : GuitarSpec. It allows to encapsulate data and to
separate generic properties of a guitar from the guitar object. Like that
we encapsulate the guitar properties out of Guitar, and put them in their
own separate object

GuitarSpec is constrcut by specificities of guitar like builder, model,
type, backwood and topwood

Guitar class change and recover only an attributes of type GuitarSpec.
A new object spec is constrcuted in constructor of Guitar

Search method in Inventory class recover the spec of guitar thanks to his
getter and can compare to differents specificities