Update code to more flexibility and support
Added numString in GuitarSpec class to store the number of strings a guitar has
because Rick wants many strings of guitar.

Added matches method in GuitarSpec class which makes the same things of
search method in Inventory class before this update. Like that, only
change this class and not in Inventory class or Guitar class.

Modified Guitar class constructor by adding attributes of GuitarSpec

Changed search method in Inventory class to delegate comparing the two
GuitarSpec objects to the GuitarSpec class, instead of handling the
comparison directly

Changed FindGuitarTest with good values in initializeInventory method.